Complete Sum All Odd Fibonacci Numbers
The goal was to get the sum of all the odd fibonacci numbers which are less than the given number. It's unnecessary to populate a set of fibonacci numbers less than the given number however, as you can simply use the += operator to continually add to a number.

As fibonacci generally goes, the previous number is always added to the last number.
You can tally a total of all the numbers by adding the new number.